Initial PoC worker able to get a USBDevice instance
This commit lays the groundwork for accessing a USB device in a worker.
Firstly, a worker is spawned.
The permission request must done in the main loop.
Once this is done the worker can access allowed USB devices using navigator.usb.getDevices()
The VID and PID is then sent to the worker to determine the correct USBDevice instance to use.