Use ensures inside of other ensures to reduce code duplication
This commit replaces the duplicate if checks inside of ensure methods with calls to other ensure methods to reduce code duplication.
This commit also adds a optional paramater to control if a ensure func sends a success message.
This paramater should be set to false when using an ensure method inside of another so that duplicate success messages are not sent.