Create handler func for worker req/res with error handling
This commit creates a centralized function, makeRequest(), which given a request, posts it and waits for a response, before doing basic error handling by checking the status code.
This allows us to reduce code duplication as we no longer have to duplicate code in each function wishing to make a worker request.